Read file exported from Learning Catalytics

This function receives as input a learning catalytics *csv* file and parse it to create a long format data frame.

read_lc(quiz_dir)
  This function only receives as input the directory where all the quizzes files are stored.

Examples

  file_to_read <- "../../datasets/Dataset1/Quiz3_session12098.csv"
  head(read_lc(file_to_read))

#> Parsed with column specification:
#> cols(
#>   .default = col_integer(),
#>   `Email address` = col_character(),
#>   `student ID` = col_character(),
#>   `Last name` = col_character(),
#>   `First name` = col_character(),
#>   `Question 2 (round 1) response` = col_character(),
#>   `Question 3 (round 1) response` = col_character(),
#>   `Question 11 (round 1) response` = col_character(),
#>   `Question 2 (round 1) score` = col_character(),
#>   `Question 1 (round 1) responded at` = col_character(),
#>   `Question 2 (round 1) responded at` = col_character(),
#>   `Question 3 (round 1) responded at` = col_character(),
#>   `Question 4 (round 1) responded at` = col_character(),
#>   `Question 5 (round 1) responded at` = col_character(),
#>   `Question 6 (round 1) responded at` = col_character(),
#>   `Question 7 (round 1) responded at` = col_character(),
#>   `Question 8 (round 1) responded at` = col_character(),
#>   `Question 9 (round 1) responded at` = col_character(),
#>   `Question 10 (round 1) responded at` = col_character(),
#>   `Question 11 (round 1) responded at` = col_character(),
#>   `Question 12 (round 1) responded at` = col_character()
#>   # ... with 2 more columns
#> )

#> See spec(...) for full column specifications.

#> Joining, by = c("email.address", "last.name", "first.name", "id", "question")

#> Joining, by = c("email.address", "last.name", "first.name", "id", "question")

#>        email.address last.name  first.name                               id
#> 1 AdoraBelle@mail.tp Dearheart Adora Belle 034e92d2de2d411ba3c2d0351ac210da
#> 2 AdoraBelle@mail.tp Dearheart Adora Belle 034e92d2de2d411ba3c2d0351ac210da
#> 3 AdoraBelle@mail.tp Dearheart Adora Belle 034e92d2de2d411ba3c2d0351ac210da
#> 4 AdoraBelle@mail.tp Dearheart Adora Belle 034e92d2de2d411ba3c2d0351ac210da
#> 5 AdoraBelle@mail.tp Dearheart Adora Belle 034e92d2de2d411ba3c2d0351ac210da
#> 6 AdoraBelle@mail.tp Dearheart Adora Belle 034e92d2de2d411ba3c2d0351ac210da
#>   question        responded.at score
#> 1        1 2009-09-03 14:52:00     1
#> 2        3 2009-09-03 14:57:00     1
#> 3        4 2009-09-03 14:59:00     1
#> 4        5 2009-09-03 15:04:00     1
#> 5        6 2009-09-03 15:07:00     1
#> 6        7 2009-09-03 15:10:00     1
#>                                             quiz n time_per_question
#> 1 ../../datasets/Dataset1/Quiz3_session12098.csv 1           NA mins
#> 2 ../../datasets/Dataset1/Quiz3_session12098.csv 2            5 mins
#> 3 ../../datasets/Dataset1/Quiz3_session12098.csv 3            2 mins
#> 4 ../../datasets/Dataset1/Quiz3_session12098.csv 4            5 mins
#> 5 ../../datasets/Dataset1/Quiz3_session12098.csv 5            3 mins
#> 6 ../../datasets/Dataset1/Quiz3_session12098.csv 6            3 mins
#>   time_per_quiz
#> 1     7006 mins
#> 2     7006 mins
#> 3     7006 mins
#> 4     7006 mins
#> 5     7006 mins
#> 6     7006 mins