Read file exported from Learning Catalytics or with the specified format.

This function receives as input a learning catalytics *csv* file and parse it to create quiz object.

read_lc(file)

Arguments

file
The name of the csv file to read

Value

it returns a quiz object containing the following columns: "email address", "question", "responded at", "score" and "quiz"

Examples

file_to_read <- "../../datasets/sample_dataset/Q01.csv"
quiz_object <- read_lc(file_to_read)

#> Parsed with column specification:
#> cols(
#>   .default = col_datetime(),
#>   `email address` = col_character(),
#>   `item01 score` = col_integer(),
#>   `item02 score` = col_integer(),
#>   `item03 score` = col_integer(),
#>   `item04 score` = col_integer(),
#>   `item05 score` = col_integer(),
#>   `item06 score` = col_integer(),
#>   `item07 score` = col_integer(),
#>   `item08 score` = col_integer(),
#>   `item09 score` = col_integer(),
#>   `item10 score` = col_integer(),
#>   `item11 score` = col_integer()
#> )

#> See spec(...) for full column specifications.

#> Joining, by = c("email address", "question")

quiz_object %>% head

#> # A tibble: 6 x 5
#>        `email address` question      `responded at` score
#>                  <chr>    <chr>              <dttm> <int>
#> 1 UTKIJ1509E@gmail.com       01 2017-02-01 06:38:20     1
#> 2 UPXWZ8813U@gmail.com       01 2017-02-01 14:32:28     0
#> 3 ECDEM7888E@gmail.com       01 2017-02-01 15:47:30     0
#> 4 NWGDV0987P@gmail.com       01 2017-02-02 07:17:36     0
#> 5 FTXHN4576J@gmail.com       01 2017-02-02 22:08:17     1
#> 6 ITNWT9882E@gmail.com       01 2017-02-03 09:10:02     0
#> # ... with 1 more variables: quiz <chr>